requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
python-dotenv==1.0.1
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
openpyxl==3.1.5
httpx==0.28.1
requests==2.32.3
paho-mqtt==2.1.0
